conlcuindo as regras de negóico de autores

+++ src/main/java/io/github/theyvison/libraryapi/service/AutorService.java
@@ -1,7 +1,9 @@
 package io.github.theyvison.libraryapi.service;
 
+import io.github.theyvison.libraryapi.exceptions.OperacaoNaoPermitidaException;
 import io.github.theyvison.libraryapi.model.Autor;
 import io.github.theyvison.libraryapi.repository.AutorRepository;
+import io.github.theyvison.libraryapi.repository.LivroRepository;
 import io.github.theyvison.libraryapi.validator.AutorValidador;
 import org.springframework.stereotype.Service;
 
@@ -13,10 +15,12 @@ import java.util.UUID;
 public class AutorService {
     private final AutorRepository autorRepository;
     private final AutorValidador autorValidador;
+    private final LivroRepository livroRepository;
 
-    public AutorService(AutorRepository autorRepository, AutorValidador autorValidador) {
+    public AutorService(AutorRepository autorRepository, AutorValidador autorValidador, LivroRepository livroRepository) {
         this.autorRepository = autorRepository;
         this.autorValidador = autorValidador;
+        this.livroRepository = livroRepository;
     }
 
     public Autor salvar(Autor autor) {
@@ -37,6 +41,9 @@ public class AutorService {
     }
 
     public void deletar(Autor autor) {
+        if (possuiLivro(autor)) {
+            throw new OperacaoNaoPermitidaException("Não é permitido excluir um autor que possui livros cadastrados!");
+        }
         autorRepository.delete(autor);
     }
 
@@ -54,5 +61,9 @@ public class AutorService {
         }
 
         return autorRepository.findAll();
+    }
+
+    public boolean possuiLivro(Autor autor) {
+        return livroRepository.existsByAutor(autor);
     }
 }
